+++ engine/include/ECS.hpp
@@ -1,16 +1,9 @@
 #pragma once
-
+#include "pch.hpp"
 #include "entity/registry.hpp"
 #include "systems/system.hpp"
 #include <spdlog/spdlog.h>
-#include "entity/snapshot.hpp"
-#include <cereal/cereal.hpp>
-#include <cereal/details/static_object.hpp>
 
-namespace std::filesystem
-{
-class path;
-}
 class System;
 
 class ECS
@@ -53,37 +46,4 @@ void ECS::AddSystem(Args&&... args)
     _systems.emplace_back(std::unique_ptr<System>(system));
 
     spdlog::info("{}, created", typeid(*system).name());
-}
-
-class EntitySerializor
-{
-public:
-    EntitySerializor(entt::registry& registry, entt::entity entity = entt::null)
-        : _registry(registry)
-        , _entity(entity)
-    {
-    }
-
-    template <class Archive>
-    void save(Archive& archive, uint32_t const version) const;
-
-private:
-    entt::registry& _registry;
-    entt::entity _entity;
-};
-
-CEREAL_CLASS_VERSION(EntitySerializor, 0);
-template <class Archive>
-void EntitySerializor::save(Archive& archive, uint32_t const version) const
-{
-    x static auto trySaveComponent = [&]<typename T>()
-    {
-        if (auto component = _registry.try_get<T>(_entity); component != nullptr)
-            archive(cereal::make_nvp(typeid(T).name(), *component));
-    };
-
-    if (version == 1)
-    {
-        trySaveComponent.template operator()<int>();
-    }
 }
